+++ Source/SwiftLintFramework/Rules/LineLengthRule.swift
@@ -20,31 +20,24 @@ public struct LineLengthRule: ParameterizedRule {
         self.parameters = parameters
     }
 
-    public let identifier = "line_length"
-
     public let parameters: [RuleParameter<Int>]
+
+    public static let description = RuleDescription(
+        identifier: "line_length",
+        name: "Line Length",
+        description: "Enforce maximum line length"
+    )
 
     public func validateFile(file: File) -> [StyleViolation] {
         return file.lines.flatMap { line in
-            for parameter in parameters.reverse() {
-                if line.content.characters.count > parameter.value {
-                    return StyleViolation(type: .Length,
-                        location: Location(file: file.path, line: line.index),
-                        severity: parameter.severity,
-                        ruleId: self.identifier,
-                        reason: "Line should be \(parameters.first!.value) characters or less: " +
-                        "currently \(line.content.characters.count) characters")
-                }
+            for param in parameters.reverse() where line.content.characters.count > param.value {
+                return StyleViolation(ruleDescription: self.dynamicType.description,
+                    severity: param.severity,
+                    location: Location(file: file.path, line: line.index),
+                    reason: "Line should be \(parameters.first!.value) characters or less: " +
+                    "currently \(line.content.characters.count) characters")
             }
             return nil
         }
     }
-
-    public let example = RuleExample(
-        ruleName: "Line Length Rule",
-        ruleDescription: "Enforce maximum line length",
-        nonTriggeringExamples: [],
-        triggeringExamples: [],
-        showExamples: false
-    )
 }