+++ Source/SwiftLintFramework/StyleViolation.swift
@@ -7,46 +7,27 @@
 //
 
 public struct StyleViolation: CustomStringConvertible, Equatable {
-    public let type: StyleViolationType
+    public let ruleDescription: RuleDescription
     public let severity: ViolationSeverity
     public let location: Location
-    public let ruleId: String
     public let reason: String?
     public var description: String {
         return XcodeReporter.generateForSingleViolation(self)
     }
 
-    public init(type: StyleViolationType, location: Location, ruleId: String,
-        reason: String? = nil) {
-        self.init(type: type, location: location, severity: .Warning, ruleId: ruleId,
-            reason: reason)
-    }
-
-    public init(type: StyleViolationType,
-        location: Location,
-        severity: ViolationSeverity,
-        ruleId: String,
-        reason: String? = nil) {
-        self.severity = severity
-        self.type = type
-        self.location = location
-        self.reason = reason
-        self.ruleId = ruleId
+    public init(ruleDescription: RuleDescription, severity: ViolationSeverity = .Warning,
+        location: Location, reason: String? = nil) {
+            self.ruleDescription = ruleDescription
+            self.severity = severity
+            self.location = location
+            self.reason = reason
     }
 }
 
 // MARK: Equatable
 
-/**
-Returns true if `lhs` StyleViolation is equal to `rhs` StyleViolation.
-
-:param: lhs StyleViolation to compare to `rhs`.
-:param: rhs StyleViolation to compare to `lhs`.
-
-:returns: True if `lhs` StyleViolation is equal to `rhs` StyleViolation.
-*/
 public func == (lhs: StyleViolation, rhs: StyleViolation) -> Bool {
-    return lhs.type == rhs.type &&
+    return lhs.ruleDescription == rhs.ruleDescription &&
         lhs.location == rhs.location &&
         lhs.severity == rhs.severity &&
         lhs.reason == rhs.reason