+++ Source/SwiftLintFrameworkTests/ConfigurationTests.swift
@@ -38,10 +38,12 @@ class ConfigurationTests: XCTestCase {
         XCTAssertEqual(disabledConfig.disabledRules,
             ["nesting", "todo"],
             "initializing Configuration with valid rules in YAML string should succeed")
-        let expectedIdentifiers = Configuration.rulesFromYAML(nil)
-            .map({ $0.identifier })
+        let expectedIdentifiers = Configuration.rulesFromYAML()
+            .map({ $0.dynamicType.description.identifier })
             .filter({ !["nesting", "todo"].contains($0) })
-        let configuredIdentifiers = disabledConfig.rules.map({ $0.identifier })
+        let configuredIdentifiers = disabledConfig.rules.map {
+            $0.dynamicType.description.identifier
+        }
         XCTAssertEqual(expectedIdentifiers, configuredIdentifiers)
 
         // Duplicate