+++ Source/SwiftLintFrameworkTests/TestHelpers.swift
@@ -14,24 +14,36 @@ func violations(string: String) -> [StyleViolation] {
     return Linter(file: File(contents: string)).styleViolations
 }
 
-private func violations(string: String, _ type: StyleViolationType) -> [StyleViolation] {
-    return violations(string).filter { $0.type == type }
+private func violations(string: String, _ description: RuleDescription) -> [StyleViolation] {
+    return violations(string).filter { $0.ruleDescription == description }
 }
 
 extension XCTestCase {
-    func verifyRule(rule: Rule,
-        type: StyleViolationType,
+    func verifyRule(ruleDescription: RuleDescription,
         commentDoesntViolate: Bool = true) {
-        let example = rule.example
-        XCTAssertEqual(example.nonTriggeringExamples.flatMap({violations($0, type)}), [])
-        XCTAssertEqual(example.triggeringExamples.flatMap({violations($0, type).map({$0.type})}),
-            Array(count: example.triggeringExamples.count, repeatedValue: type))
+        XCTAssertEqual(
+            ruleDescription.nonTriggeringExamples.flatMap({violations($0, ruleDescription)}),
+            []
+        )
+        XCTAssertEqual(
+            ruleDescription.triggeringExamples.flatMap({
+                violations($0, ruleDescription).map({$0.ruleDescription})
+            }),
+            Array(count: ruleDescription.triggeringExamples.count, repeatedValue: ruleDescription))
 
         if commentDoesntViolate {
-            XCTAssertEqual(example.triggeringExamples.flatMap({violations("/** " + $0, type)}), [])
+            XCTAssertEqual(
+                ruleDescription.triggeringExamples.flatMap({
+                    violations("/** " + $0, ruleDescription)
+                }),
+                []
+            )
         }
 
-        let command = "// swiftlint:disable \(rule.identifier)\n"
-        XCTAssertEqual(example.triggeringExamples.flatMap({violations(command + $0, type)}), [])
+        let command = "// swiftlint:disable \(ruleDescription.identifier)\n"
+        XCTAssertEqual(
+            ruleDescription.triggeringExamples.flatMap({violations(command + $0, ruleDescription)}),
+            []
+        )
     }
 }